fix ld64 to compile with libstdc++ fix automake warnings

+++ cctools/ld64/src/ld/passes/branch_island.cpp
@@ -498,7 +498,7 @@ static void makeIslandsForSection(const Options& opts, ld::Internal& state, ld::
 				if ( target->section().type() == ld::Section::typeStub )
 					dstAddr = totalTextSize;
 				int64_t displacement = dstAddr - srcAddr;
-				TargetAndOffset finalTargetAndOffset = { target, addend };
+				TargetAndOffset finalTargetAndOffset = { target, static_cast<uint32_t>(addend) };
 				const int64_t kBranchLimit = kBetweenRegions;
 				if ( crossSectionBranch && ((displacement > kBranchLimit) || (displacement < (-kBranchLimit))) ) {
 					const ld::Atom* island;
